split out old tests w/ diagnostics
