+++ test/runtests.jl
@@ -71,7 +71,7 @@ function make_test_data(rng, sample_rate, samples_per_record, n_records, ::Type{
                      :ptaf => [17])
 end
 
-include("test_edf_to_samples_info.in")
+# include("test_edf_to_samples_info.in")
 
 @testset "OndaEDF" begin
     include("signal_labels.jl")
